+++ lang/en/theme_foundation.php
@@ -119,6 +119,21 @@ $string['alerttextdesc'] = 'Text for alert {$a->number}.';
 $string['alertpage'] = 'Alert {$a->number} page';
 $string['alertpagedesc'] = 'Show alert {$a->number} on the given page or \'All\'.';
 
+$string['brandsheading'] = 'Brand icons settings';
+$string['brandsheadingdesc'] = 'Set the brand icon settings.  The brand icon \'name\' is  - this is the FontAwesome name without the prefixing \'fa-\' so \'fa-train\' simply becomes \'train\'.  Depending on the \'fav\' setting to enable use of FontAwsome 5, please look at \'fontawesome.com/icons?d=gallery&s=brands&m=free\' if set and \'fontawesome.com/v4.7.0/icons/#brand\' if not, for the name of the icon.  The theme will do the rest with the class names.';
+
+$string['numberofbrands'] = 'Brands';
+$string['numberofbrandsdesc'] = 'Number of brands between {$a->lower} and {$a->upper}.  After changing and \'Saving changes\', refresh the page.';
+
+$string['brandsettingheading'] = 'Brand {$a->number} settings';
+
+$string['enablebrand'] = 'Brand {$a->number} enable';
+$string['enablebranddesc'] = 'Enable or disable brand {$a->number}.';
+$string['brandiconname'] = 'Brand icon {$a->number} name';
+$string['brandiconnamedesc'] = 'Icon name for brand {$a->number}.';
+$string['brandiconurl'] = 'Brand icon {$a->number} URL';
+$string['brandiconurldesc'] = 'URL for brand icon {$a->number}.';
+
 $string['loginbackgroundheading'] = 'Login background image settings';
 $string['loginbackgroundheadingdesc'] = 'Set the login background image settings.';
 $string['loginbackground'] = 'Login background image';